chore: move backend files to api folder
