+++ api/src/global.d.ts
