+++ api/src/routes/health-check/validation.ts
