+++ api/src/utils/auth.ts
